requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.9.1
pwdlib[argon2]==0.2.1
pyjwt==2.10.1
python-dotenv==1.1.0